apps/services/llm/requirements.txt
fastapi
uvicorn
psycopg2-binary
qdrant-client
langchain-ollama
ollama
pydantic
python-multipart
